requirements.txt
pandas>=2.2
openpyxl>=3.1
xlsxwriter>=3.2
numpy>=2.0
pyyaml>=6.0
pydantic>=2.10
python-dotenv>=1.0
httpx>=0.27
tenacity>=9.0
tqdm>=4.66
sentence-transformers>=3.0
rank-bm25>=0.2.2
transformers>=4.44
torch>=2.3
datasets>=3.0
ragas>=0.2.12
langchain-google-genai>=2.0
langchain-ollama>=0.2
google-genai>=1.0
google-generativeai>=0.8
pytest>=8.3
